feat(nvmx): abstract block devices for nexus children (#807)
Nexus now treats all its replica devices as abstract block devices,
which allows clear separation, and uses abstract block device I/O
handles for serving child I/O.

In order to prevent start of controller resets in response to I/O errors
(mostly connection failures), maximum number of controller resets
introduced along with reset cool down interval.

Per-channel I/O operation accounting is introduced to have better I/O
observability/statistics.

Implements CAS-781

+++ mayastor/src/bdev/nexus/nexus_metadata.rs
@@ -708,7 +708,10 @@ impl NexusChild {
                 && partition.ent_name.name == "MayaMeta"
             {
                 let mut metadata = NexusMetaData {
-                    header: MetaDataHeader::new(bdev.block_len(), &partition),
+                    header: MetaDataHeader::new(
+                        bdev.block_len() as u32,
+                        &partition,
+                    ),
                     index: Vec::new(),
                 };
                 self.sync_metadata(&mut metadata).await?;
